backend/requirements.txt
cryptography
Flask
Flask-SQLAlchemy
flask-cors
PyMySQL
PyPDF2
pandas
openpyxl
python-slugify
sqlalchemy-serializer
cerberus